refactor routes, home page, add pagination and search options

+++ server/controllers/posts.js
@@ -17,7 +17,12 @@ export const getPosts = (req, res) => {
 export const createPost = (req, res) => {
   const post = req.body;
 
-  const newPost = new PostMessage(post);
+  if(!req.userId)
+    return res.status(401).json({ message: 'Unathenticated' });
+
+  const newPost = new PostMessage({...post, creator_id: req.userId});
+
+  console.log(newPost.createdAt);
 
   newPost.save()
     .then((data) => res.status(201).json(data))
@@ -31,8 +36,8 @@ export const likePost = (req, res) => {
   if(!mongoose.Types.ObjectId.isValid(id))
     return res.status(404).send('No post with that id');
 
-  // if(!req.id)
-  //   return res.status(401).json({ message: 'Unathenticated' });
+  if(!req.userId)
+    return res.status(401).json({ message: 'Unathenticated' });
 
   PostMessage.findByIdAndUpdate({_id: id}, { likes: post.likes }, {new: true})
     .then((data) => res.status(201).json(data))
@@ -46,6 +51,9 @@ export const updatePost = (req, res) => {
   if(!mongoose.Types.ObjectId.isValid(id))
     return res.status(404).send('No post with that id');
 
+  if(!req.userId)
+    return res.status(401).json({ message: 'Unathenticated' });
+
   PostMessage.findByIdAndUpdate({_id: id}, post, {new: true})
     .then((data) => res.status(201).json(data))
     .catch((err) => res.status(500).json({error: err.message}));
@@ -56,6 +64,9 @@ export const deletePost = (req, res) => {
 
   if(!mongoose.Types.ObjectId.isValid(id))
     return res.status(404).send('No post with that id');
+
+  if(!req.userId)
+    return res.status(401).json({ message: 'Unathenticated' });
 
   PostMessage.findByIdAndDelete(id)
     .then((data) => res.status(200).json(data))
